Scaffold about page, about/ ui components, and move landing page into home/
