+++ app/page.tsx
@@ -1,4 +1,4 @@
-import Hero from '@/app/ui/about/hero'
+import Hero from '@/app/ui/home/hero'
 
 export default function Home() {
     return (
