+++ app/ui/about/features.tsx
@@ -1,0 +1,12 @@
+export default function Features() {
+    return (
+        <section className="bg-page">
+            <div className="mx-auto max-w-3xl px-4 py-12">
+                <h1 className="text-2xl font-bold tracking-tight text-subheading">
+                    Features
+                </h1>
+                <div>Features</div>
+            </div>
+        </section>
+    )
+}
