+++ app/ui/about/overview.tsx
@@ -1,0 +1,12 @@
+export default function Overview() {
+    return (
+        <section className="bg-page">
+            <div className="mx-auto max-w-3xl px-4 py-12">
+                <h1 className="text-3xl font-bold tracking-tight text-heading">
+                    Overview
+                </h1>
+                <div>Overview</div>
+            </div>
+        </section>
+    )
+}
