+++ app/ui/about/roadmap.tsx
@@ -1,0 +1,12 @@
+export default function Roadmap() {
+    return (
+        <section className="bg-page">
+            <div className="mx-auto max-w-3xl px-4 py-12">
+                <h1 className="text-2xl font-bold tracking-tight text-subheading">
+                    Roadmap
+                </h1>
+                <div>Roadmap</div>
+            </div>
+        </section>
+    )
+}
